fix and refactor: fix imports, and change basic folders, update main.py
Fix imports, change models and schemas to basic_models and basic_schemas, add main api router to FastAPI app.
