+++ backend/app/api/v1/articles/models.py
@@ -8,8 +8,8 @@ with its associated properties.
 
 from tortoise import fields
 
-from app.models.base import AbstractBaseModel
-from app.models.timestamp import TimestampMixin
+from basic_models.base import AbstractBaseModel
+from basic_models.timestamp import TimestampMixin
 
 
 class Article(AbstractBaseModel, TimestampMixin):
